fix: drag coordinate offset + memory optimization
Drag coordinate fix (root cause of 'appears at right side' and
'offset cannot update'):
- Widget::GetPaintOffset() virtual method returns paint-time transform
  (default zero). Draggable overrides to return drag_offset_.
- Application::DispatchPointerEvent/Move subtract GetPaintOffset() when
  computing local_pos, converting visual coords to widget-local coords.
- Draggable::OnPointerDown records start_offset_ (current drag_offset_)
  and press_pos_; OnPointerMove computes absolute offset as
  start_offset_ + (current_pos - press_pos). This fixes both the
  initial jump (widget appearing at cursor side) and stale offset.
- Widget::HitTest remains virtual; Draggable::HitTest subtracts
  drag_offset_ so clicks at visual position hit correctly.

Memory optimization:
- Application::window_width_/height_: int -> std::uint16_t
- RenderLayer::window_width_/height_: int -> std::uint16_t
- RenderLayer members: in-class default initializers
- .clang-tidy: disable modernize-use-equals-default (false positive on
  non-trivial constructors)

clang-tidy: 0 code warnings

+++ src/render/render_layer.cpp
@@ -53,8 +53,8 @@ bool RenderLayer::Start(int width, int height, std::string_view title,
     return false;
   }
 
-  window_width_ = width;
-  window_height_ = height;
+  window_width_ = static_cast<std::uint16_t>(width);
+  window_height_ = static_cast<std::uint16_t>(height);
 
   LOG(INFO) << "RenderLayer starting: " << width << "x" << height
             << " backend=" << FLAGS_render_backend;
@@ -215,8 +215,8 @@ void RenderLayer::GetWindowSize(int& width, int& height) const noexcept {
     height = platform_bridge_->GetHeight();
     return;
   }
-  width = window_width_;
-  height = window_height_;
+  width = static_cast<int>(window_width_);
+  height = static_cast<int>(window_height_);
 }
 
 void RenderLayer::RenderLoop() {
